test(interpreter): split fd test files under the 300-line guideline
`fd-table.test.ts` (323) and `numeric-fds.input.test.ts` (364) had grown
past the repo's 300-line limit for test files. Split by theme, no
assertions changed:

- fd-table.encoding.test.ts   — encoding round-trips, content vs. marker
- fd-table.access.test.ts     — accessors, read/advance, snapshot/restore
- numeric-fds.marker-content.test.ts — marker-shaped file content, lifted
  out of numeric-fds.input.test.ts
